+++ lib/domain/menstruation/menstruation_store.dart
@@ -8,7 +8,7 @@ import 'package:pilll/domain/menstruation/menstruation_history_card_state.dart';
 import 'package:pilll/entity/menstruation.codegen.dart';
 import 'package:pilll/native/health_care.dart';
 import 'package:pilll/database/diary.dart';
-import 'package:pilll/service/menstruation.dart';
+import 'package:pilll/database/menstruation.dart';
 import 'package:pilll/service/pill_sheet.dart';
 import 'package:pilll/service/pill_sheet_group.dart';
 import 'package:pilll/service/setting.dart';
