+++ lib/inquiry/inquiry.dart
@@ -3,7 +3,7 @@ import 'package:pilll/entity/pill_sheet_group.codegen.dart';
 import 'package:pilll/database/database.dart';
 import 'package:pilll/entity/menstruation.codegen.dart';
 import 'package:pilll/entity/setting.codegen.dart';
-import 'package:pilll/service/menstruation.dart';
+import 'package:pilll/database/menstruation.dart';
 import 'package:pilll/service/pill_sheet_group.dart';
 import 'package:pilll/service/setting.dart';
 import 'package:pilll/util/environment.dart';
